requirements.txt
garminconnect>=0.2.22
caldav>=1.3.9
icalendar>=5.0.13
Mastodon.py>=1.8.1
gpxpy>=1.6.2
staticmap>=0.5.4
Pillow>=10.0.0
requests>=2.31.0
tomllib; python_version < "3.11"
